Refactor of main.ts
