+++ index.html
@@ -64,6 +64,6 @@
       </div>
     </div>
     <div id="blocklyDiv" style="position: absolute"></div>
-    <script type="module" src="src/main.ts"></script>
+    <script type="module" src="src/index.ts"></script>
   </body>
 </html>
